Finish component BurgerIngredients
Need to rectify screen size
