+++ src/css/components/app-header.css
@@ -9,7 +9,6 @@
     right: 0%;
     top: 0px;
     background: #1C1C21;
-    box-shadow: 0px 24px 32px rgba(0, 0, 0, 0.04), 0px 16px 24px rgba(0, 0, 0, 0.04), 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 0px 1px rgba(0, 0, 0, 0.04);
 }
 
 .app-header_navigation__panel-container {
