+++ src/css/components/burger-card.css
@@ -1,14 +1,25 @@
 .burger-card {
+    position: relative;
     display: flex;
     flex-direction: column;
     align-items: center;
     padding: 0px;
+    margin-top: 24px;
     gap: 8px;
     width: 272px;
     height: 208px;
     flex: none;
     order: 0;
-    flex-grow: 1;
+}
+
+.burger-card__counter::after {
+    position: absolute;
+    left: 0%;
+    right: 0%;
+    top: 0%;
+    bottom: 0%;
+    background: #4C4CFF;
+    border-radius: 16px;
 }
 
 .burger-card__img {
@@ -34,7 +45,7 @@
 }
 
 .burger-card__name {
-    width: 272px;
+    width: 100%;
     height: 48px;
     background: #131316;
     flex: none;
