+++ src/css/index.css
@@ -1,3 +1,18 @@
+::-webkit-scrollbar {
+  width: 8px; /* ширина для вертикального скролла */
+  background-color: #2F2F37;;
+}
+
+/* ползунок скроллбара */
+::-webkit-scrollbar-thumb {
+  background-color: #8585AD;
+  height: 492px;
+}
+
+::-webkit-scrollbar-thumb:hover {
+  background-color: #2F2F37;;
+}
+
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
@@ -7,7 +22,7 @@ body {
   -moz-osx-font-smoothing: grayscale;
   position: relative;
   width: 100%;
-  height: 1000px;
+  height: 100vh;
   background: #131316;
 }
 
@@ -19,7 +34,7 @@ code {
 .wrapper {
   position: relative;
   width: 1280px;
-  height: 100%;
+  height: 100vh;
   margin: 0 auto;
   border-radius: 10px;
   max-width: 1400px;
