+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from './App.js'
-import './index.css'
+import App from './components/App.js'
+import './css/index.css'
 
 
 const root = ReactDOM.createRoot(
